# ESM-ui/src/app/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Yehor",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Term of service",
    "contact": {
      "name": "Yehor",
      "url": "https://t.me/dead_1y",
      "email": "kulish233@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://google.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888/api/v1",
      "description": "Local Env"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/requests/respond-request": {
      "put": {
        "tags": ["Requests"],
        "operationId": "respondRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/set-new-password": {
      "post": {
        "tags": ["User"],
        "operationId": "setNewPasswordForUserByAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/services/create": {
      "post": {
        "tags": ["Services"],
        "operationId": "createService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/requests/create-request": {
      "post": {
        "tags": ["Requests"],
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleRequests"
                }
              }
            }
          }
        }
      }
    },
    "/employees/create": {
      "post": {
        "tags": ["Employees"],
        "operationId": "createEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["User"],
        "operationId": "getAllUsersWithDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["User"],
        "operationId": "findUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["User"],
        "operationId": "getMyUserDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/services": {
      "get": {
        "tags": ["Services"],
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Services"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": ["Services"],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "operationId": "findAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RolesWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/roles/me": {
      "get": {
        "tags": ["Roles"],
        "operationId": "findMyRoles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RolesWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests": {
      "get": {
        "tags": ["Requests"],
        "operationId": "findAll_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRoleRequestsWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests/{id}": {
      "get": {
        "tags": ["Requests"],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleRequestsWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/requests/authorized-requests": {
      "get": {
        "tags": ["Requests"],
        "operationId": "findRequestsThatCanBeResponded",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRoleRequestsWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/positions": {
      "get": {
        "tags": ["Positions"],
        "operationId": "findAll_3",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Positions"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees": {
      "get": {
        "tags": ["Employees"],
        "operationId": "findAll_4",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeWrapper"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees/no-role": {
      "get": {
        "tags": ["Employees"],
        "operationId": "findEmployeesWithoutUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employees"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/departments": {
      "get": {
        "tags": ["Department"],
        "operationId": "findAll_5",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Departments"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/count": {
      "get": {
        "tags": ["Dashboard"],
        "operationId": "getCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardCountResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateRequestBody": {
        "required": ["id", "response"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "response": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "ForgotPasswordRequest": {
        "required": ["password", "username"],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ServiceRequest": {
        "required": ["serviceTitle"],
        "type": "object",
        "properties": {
          "serviceTitle": {
            "type": "string"
          }
        }
      },
      "UserRoleRequestRequest": {
        "required": ["requestedServiceName"],
        "type": "object",
        "properties": {
          "requestedServiceName": {
            "type": "string"
          },
          "owner": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "Departments": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "depTitle": {
            "type": "string"
          }
        }
      },
      "Employees": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Departments"
          },
          "position": {
            "$ref": "#/components/schemas/Positions"
          }
        }
      },
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "Positions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "posTitle": {
            "type": "string"
          }
        }
      },
      "Roles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string"
          },
          "roleDesc": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/Services"
          },
          "owner": {
            "type": "boolean"
          }
        }
      },
      "Services": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "srvTitle": {
            "type": "string"
          }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "UserRoleRequests": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userRoleRequestedBy": {
            "$ref": "#/components/schemas/Users"
          },
          "requestedRole": {
            "$ref": "#/components/schemas/Roles"
          },
          "requestedService": {
            "$ref": "#/components/schemas/Services"
          },
          "note": {
            "type": "string"
          },
          "roleThatCanRespond": {
            "$ref": "#/components/schemas/Roles"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedByUser": {
            "$ref": "#/components/schemas/Users"
          }
        }
      },
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Roles"
            }
          },
          "empId": {
            "$ref": "#/components/schemas/Employees"
          },
          "name": {
            "type": "string"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
      "EmployeeRequest": {
        "required": ["depId", "fullName", "posId"],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "posId": {
            "type": "integer",
            "format": "int32"
          },
          "depId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RegistrationRequest": {
        "required": ["empId", "password", "username"],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "empId": {
            "type": "integer",
            "format": "int32"
          },
          "password": {
            "maxLength": 30,
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": ["password", "username"],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "maxLength": 30,
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "UserWrapper": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "posTitle": {
            "type": "string"
          },
          "depTitle": {
            "type": "string"
          },
          "roleNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "srvNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ServiceWrapper": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "integer",
            "format": "int32"
          },
          "serviceTitle": {
            "type": "string"
          },
          "ownerRole": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "ownerUsers": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "notOwnerRoleWithAccess": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
      "RolesWrapper": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string"
          },
          "roleDesc": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "owner": {
            "type": "boolean"
          }
        }
      },
      "UserRoleRequestsWrapper": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string"
          },
          "lastModifiedBy": {
            "type": "string"
          },
          "requestedRole": {
            "type": "string"
          },
          "requestedService": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string"
          },
          "userRoleRequestedBy": {
            "type": "string"
          },
          "roleThatCanRespond": {
            "type": "string"
          }
        }
      },
      "EmployeeWrapper": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "departmentTitle": {
            "type": "string"
          },
          "positionTitle": {
            "type": "string"
          }
        }
      },
      "DashboardCountResponse": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "integer",
            "format": "int32"
          },
          "services": {
            "type": "integer",
            "format": "int32"
          },
          "requests": {
            "type": "integer",
            "format": "int32"
          },
          "waitingForResponse": {
            "type": "integer",
            "format": "int32"
          },
          "users": {
            "type": "integer",
            "format": "int32"
          },
          "employees": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}