# ESM-ui/src/app/services/models/dashboard-count-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DashboardCountResponse {
  employees?: number;
  requests?: number;
  roles?: number;
  services?: number;
  users?: number;
  waitingForResponse?: number;
}