# ESM-ui/src/app/services/models/user-role-requests-wrapper.ts
/* tslint:disable */
/* eslint-disable */
export interface UserRoleRequestsWrapper {
  createdDate?: string;
  currentStatus?: string;
  id?: number;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  note?: string;
  requestedRole?: string;
  requestedService?: string;
  roleThatCanRespond?: string;
  userRoleRequestedBy?: string;
}