# ESM-ui/src/app/services/models/user-wrapper.ts
/* tslint:disable */
/* eslint-disable */
export interface UserWrapper {
  depTitle?: string;
  fullName?: string;
  id?: number;
  posTitle?: string;
  roleNames?: Array<string>;
  srvNames?: Array<string>;
  username?: string;
}